Update UI for new API and add node effect
